added basic implement for DTO and repo functions for fetching data

+++ src/main/java/com/alex/barselplus_backend/Patient.java
@@ -30,7 +30,7 @@ public class Patient {
     private LocalDate dateOfBirth;
 
     @Column(name = "national_ID", unique = true, nullable = false)
-    private Integer nationalID;
+    private Long nationalID;
 
     @Column(name = "address")
     private String address;
